refactor: arranged directory for separation of concerns
