+++ src/components/assets/GenericMale.png
