+++ src/public/icons/512x512.png
