+++ src/public/sw.js
@@ -1,0 +1,1 @@
+if(!self.define){let e,s={};const i=(i,c)=>(i=new URL(i+".js",c).href,s[i]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=s,document.head.appendChild(e)}else e=i,importScripts(i),s()})).then((()=>{let e=s[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e})));self.define=(c,n)=>{const a=e||("document"in self?document.currentScript.src:"")||location.href;if(s[a])return;let r={};const t=e=>i(e,a),d={module:{uri:a},exports:r,require:t};s[a]=Promise.all(c.map((e=>d[e]||t(e)))).then((e=>(n(...e),r)))}}define(["./workbox-f77a352b"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/551.d42a0d5390997d49.js",revision:"d42a0d5390997d49"},{url:"/_next/static/chunks/714-1daed6568b99ce81.js",revision:"1daed6568b99ce81"},{url:"/_next/static/chunks/framework-2c79e2a64abdb08b.js",revision:"2c79e2a64abdb08b"},{url:"/_next/static/chunks/main-505fb84c7ea07d72.js",revision:"505fb84c7ea07d72"},{url:"/_next/static/chunks/pages/_app-87ba9f4ee8f67e52.js",revision:"87ba9f4ee8f67e52"},{url:"/_next/static/chunks/pages/_error-8353112a01355ec2.js",revision:"8353112a01355ec2"},{url:"/_next/static/chunks/pages/index-db11b2bb0b5d7dec.js",revision:"db11b2bb0b5d7dec"},{url:"/_next/static/chunks/polyfills-c67a75d1b6f99dc8.js",revision:"837c0df77fd5009c9e46d446188ecfd0"},{url:"/_next/static/chunks/webpack-624128614328ee73.js",revision:"624128614328ee73"},{url:"/_next/static/css/62a4007fa89793ec.css",revision:"62a4007fa89793ec"},{url:"/_next/static/media/GenericFemale.40cafc87.png",revision:"c49a972fc128aa358f6e280422f56207"},{url:"/_next/static/media/GenericMale.592f9e48.png",revision:"5145708211d1de9ae5652d9abac36850"},{url:"/_next/static/nktMhqU2p-C9mx6h8V40w/_buildManifest.js",revision:"1b531b6c6f74e77070a1e08887aa601e"},{url:"/_next/static/nktMhqU2p-C9mx6h8V40w/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/cslogo.png",revision:"7bd7761cc6b68d7cb2d281497633466b"},{url:"/favicon-16x16.png",revision:"84acc9f3836db2092c91ffa1bb5665da"},{url:"/favicon-32x32.png",revision:"fa221cd188a53aca4f9f3d85c88043e0"},{url:"/favicon.ico",revision:"d74230eb8af2a7a4800427e6b4a06658"},{url:"/icons/16x16.png",revision:"35d75eeaca35d461d32db4aa044fdc45"},{url:"/icons/192x192.png",revision:"433ff59bce46f58091b316d56888ca6b"},{url:"/icons/32x32.png",revision:"4b560241068bc5edfedd5d42538da6f3"},{url:"/icons/384x384.png",revision:"c541845ff2917c198364acc7347562d3"},{url:"/icons/512x512.png",revision:"7b829d638502aa420163d3a0f727ffb5"},{url:"/icons/PCLUB Logo.png",revision:"bee032bfa91c5b06edb8c0b280879461"},{url:"/icons/apple-touch-icon.png",revision:"ebf3e76a5c84b481ccb31c2aaa784518"},{url:"/manifest.json",revision:"7d8ef999d7aa6521c712d5bb8c8cab42"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:i,state:c})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET")}));
